+++ tests/test_copy.c
@@ -34,8 +34,8 @@ static char* test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, i
             storage.properties.blosc.filename = filename;
             storage.properties.blosc.enforceframe = enforceframe;
             for (int i = 0; i < ndim; ++i) {
-                storage.properties.blosc.chunkshape[i] = chunkshape[i];
-                storage.properties.blosc.blockshape[i] = blockshape[i];
+                storage.properties.blosc.chunkshape[i] = (uint32_t) chunkshape[i];
+                storage.properties.blosc.blockshape[i] = (uint32_t) blockshape[i];
             }
             storage.properties.blosc.nmetalayers = 1;
             storage.properties.blosc.metalayers[0].name = "random";
@@ -52,8 +52,8 @@ static char* test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, i
     for (int i = 0; i < ndim; ++i) {
         buffersize *= shape[i];
     }
-    uint8_t *buffer = malloc(buffersize);
-    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, buffersize / itemsize));
+    uint8_t *buffer = malloc((size_t) buffersize);
+    MU_ASSERT("Buffer filled incorrectly", fill_buf(buffer, itemsize, (size_t) (buffersize / itemsize)));
 
     /* Create caterva_array_t with original data */
     caterva_array_t *src;
@@ -83,8 +83,8 @@ static char* test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, i
             storage2.properties.blosc.filename = filename2;
             storage2.properties.blosc.enforceframe = enforceframe2;
             for (int i = 0; i < ndim; ++i) {
-                storage2.properties.blosc.chunkshape[i] = chunkshape2[i];
-                storage2.properties.blosc.blockshape[i] = blockshape2[i];
+                storage2.properties.blosc.chunkshape[i] = (uint32_t) chunkshape2[i];
+                storage2.properties.blosc.blockshape[i] = (uint32_t) blockshape2[i];
             }
             break;
         default:
@@ -94,7 +94,7 @@ static char* test_copy(caterva_context_t *ctx, uint8_t itemsize, uint8_t ndim, i
     caterva_array_t *dest;
     MU_ASSERT_CATERVA(caterva_array_copy(ctx, src, &storage2, &dest));
 
-    uint8_t *buffer_dest = malloc(buffersize);
+    uint8_t *buffer_dest = malloc((size_t) buffersize);
     MU_ASSERT_CATERVA(caterva_array_to_buffer(ctx, dest, buffer_dest, buffersize));
 
     /* Testing */
